pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG Chatbot for Published Book"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi==0.104.1",
    "langchain==0.0.352",
    "cohere==4.3.1",
    "qdrant-client==1.8.0",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "python-dotenv==1.0.0",
    "tiktoken==0.5.2",
    "pypdf==3.17.4"
]

[project.optional-dependencies]
dev = [
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1"
]

[tool.setuptools.packages.find]
where = ["backend/src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true